Implemented new mechanics to the specifications required by interviewer
Restructured project code to use raycast and planes vs the Unity UI system. This was achieved by refactoring all code that used the UI system and button clicks to instead use the new "RaycastHandler" script. This script using colliders and listeners in the update method to achieve functionality. I also refactored the "ChangeDenomination" script to use a custom array that can be changed in editor to set the current denomination. These mechanics are functional but will be refined further in future commits.